refactor: change details-page directory
