+++ class-components/src/pages/search-page/details-page/details-page.tsx
