+++ class-components/src/routes/router.tsx
@@ -1,9 +1,9 @@
 import { createBrowserRouter } from 'react-router-dom';
 
 import { Layout } from '@/components/layout/layout';
-import { DetailsPage } from '@/pages/details-page';
-import { loader as DetailsLoader } from '@/pages/details-page-loader';
 import { NotFoundPage } from '@/pages/not-found-page/not-found-page';
+import { DetailsPage } from '@/pages/search-page/details-page/details-page';
+import { loader as DetailsLoader } from '@/pages/search-page/details-page/details-page-loader';
 
 export const APP_ROUTES = {
   cards: '/cards',
